Updated the metric persistant

+++ src/main/java/com/zaga/entity/queryentity/metrics/MetricDTO.java
@@ -17,8 +17,8 @@ import lombok.NoArgsConstructor;
 @MongoEntity(collection = "MetricDTO", database = "OtelMetric")
 public class MetricDTO {
     private LocalDateTime date;
-    private Long cpuUsage;
-    private Long memoryUsage;
+    private Double cpuUsage;
+    private Integer memoryUsage;
     private String serviceName;
     
 }
